feat: add LoginValidation
include
1. CheckLogin.js => bring input data to find account data
2. FetchData.js => get account data from account.json
3. Token => create token and check token
